handler.py
import runpod
from ultralytics import YOLO
import torch
import cv2
import numpy as np
from PIL import Image
import base64
import io

# 1. Models ko GLOBAL load karna hai (taake har request par load na hon)
print("🧠 Loading All Models...")
m0_model = torch.load("models/M0.pth", map_location="cpu").eval() # ResNet50 logic yahan check karna
m1_model = YOLO("models/best_m1.pt")
m2_model = YOLO("models/best_m2.pt")
m3_model = YOLO("models/best_m3.pt")

def handler(event):
    # 'event' me frontend se bheji gayi image ayegi
    input_data = event.get("input")
    image_b64 = input_data.get("image") # Base64 string
    
    # Base64 string ko OpenCV image me convert karna
    img_bytes = base64.b64decode(image_b64)
    img_nparr = np.frombuffer(img_bytes, np.uint8)
    img_bgr = cv2.imdecode(img_nparr, cv2.IMREAD_COLOR)
    
    # --- YAHAN TUMHARI PURANI DETECTION LOGIC AYEGI ---
    # (Slicing, NMS, Severity calculation waghera jo humne app.py me ki thi)
    
    # Final Result JSON ki surat me return karna hai
    report = {
        "status": 200,
        "data": {
            "metadata": {"skin_type": "Oily"}, # Example
            "results": [] # Pure analysis results
        }
    }
    
    return report

# Runpod worker start karna
runpod.serverless.start({"handler": handler})