requirements.txt
ultralytics
opencv-python-headless
torchvision
pillow
runpod